Remove Joda Time and use java.time package instead
Joda Time has been superseded by the Java 8 java.time package
which has more or less the same API. This commit removes all
usage of Joda Time and replaces it with the java.time classes.
This is done so that we can remove the dependency on Joda Time.

+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/compose/component/Scaffolding.kt
@@ -78,7 +78,7 @@ fun ScaffoldWithTopBarAndDeviceName(
     isIconAndLogoVisible: Boolean = true,
     snackbarHostState: SnackbarHostState = remember { SnackbarHostState() },
     deviceName: String?,
-    timeLeft: Int?,
+    timeLeft: Long?,
     content: @Composable (PaddingValues) -> Unit,
 ) {
     Scaffold(
